refactor: override static property in child class
