+++ src/js/material.ts
@@ -1,13 +1,6 @@
 import { Standata } from "./base";
-import MATERIALS from "./entities/materials.json";
+import MATERIALS from "./runtime_data/materials.json";
 
 export class MaterialStandata extends Standata {
-    constructor() {
-        super(MATERIALS);
-    }
-
-    // eslint-disable-next-line class-methods-use-this
-    protected loadEntity(filename: string): object | undefined {
-        return MATERIALS[filename];
-    }
+    static runtimeData = MATERIALS;
 }
