+++ src/js/workflow.ts
@@ -1,13 +1,6 @@
 import { Standata } from "./base";
-import WORKFLOWS from "./entities/workflows.json";
+import WORKFLOWS from "./runtime_data/workflows.json";
 
 export class WorkflowStandata extends Standata {
-    constructor() {
-        super(WORKFLOWS);
-    }
-
-    // eslint-disable-next-line class-methods-use-this
-    protected loadEntity(filename: string): object | undefined {
-        return WORKFLOWS[filename];
-    }
+    static runtimeData = WORKFLOWS;
 }
